Remove mentions of numerical formatting options
By using %s everywhere we get consistency and reduce the things that the
students have to pick up at the beginning.

I have noticed that the exercise with the floating-point precision and
column markup is a massive slowdown for most students who spend a long
time copying out by-hand a very syntax-heavy code block.

Nowhere in the course (except one place where it can be explained
simply) do we use any of the string fomrmatting features anyway.